+++ astropy/table/tests/test_row.py
@@ -279,3 +279,32 @@ def test_row_tuple_column_slice_transaction():
         t[1]['a', 'b'] = (-1, -1 * u.s)  # Bad unit
     assert "'s' (time) and 'm' (length) are not convertible" in str(err)
     assert t[1] == tc[1]
+
+
+def test_uint_indexing():
+    """
+    Test that accessing a row with an unsigned integer
+    works as with a signed integer.  Similarly tests
+    that printing such a row works.
+
+    This is non-trivial: adding a signed and unsigned
+    integer in numpy results in a float, which is an
+    invalid slice index.
+
+    Regression test for gh-7464.
+    """
+    t = table.Table([[1., 2., 3.]], names='a')
+    assert t['a'][1] == 2.
+    assert t['a'][np.int(1)] == 2.
+    assert t['a'][np.uint(1)] == 2.
+    assert t[np.uint(1)]['a'] == 2.
+
+    trepr = ['<Row index=1>',
+             '   a   ',
+             'float64',
+             '-------',
+             '    2.0']
+
+    assert repr(t[1]).splitlines() == trepr
+    assert repr(t[np.int(1)]).splitlines() == trepr
+    assert repr(t[np.uint(1)]).splitlines() == trepr